Clean up some things
