+++ lib/shared/widgets/screen_wrapper.dart
@@ -40,7 +40,8 @@ class _ScreenWrapper extends State<ScreenWrapper> {
       appBar: widget.appBarWidget,
       bottomNavigationBar: widget.bottomNavigationBarWidget,
       body: Container(
-          padding: widget.bodyPadding ?? const EdgeInsets.all(16),
+          padding:
+              widget.bodyPadding ?? const EdgeInsets.only(left: 16, right: 16),
           width: MediaQuery.of(context).size.width,
           height: MediaQuery.of(context).size.height,
           child: widget.children[0]),
