+++ lib/utils/shell.dart
@@ -4,7 +4,7 @@ import 'package:provider/provider.dart';
 import 'package:receipt_wrangler_mobile/groups/nav/group_ui_shell_builder.dart';
 import 'package:receipt_wrangler_mobile/models/app_bar_model.dart';
 import 'package:receipt_wrangler_mobile/models/bottom_nav_model.dart';
-import 'package:receipt_wrangler_mobile/receipts/nav/recdipts-ui-shell-builder.dart';
+import 'package:receipt_wrangler_mobile/receipts/nav/receipts_ui_shell_builder.dart';
 
 void handleShellUIUpdate(BuildContext context, {String? fullPath}) {
   var router = GoRouter.of(context);
